# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InterviewMessageRole =
  (typeof InterviewMessageRole)[keyof typeof InterviewMessageRole];

export const InterviewMessageRole = {
  student: "student",
  advisor: "advisor",
} as const;

export interface InterviewMessage {
  role: InterviewMessageRole;
  content: string;
}

export interface InterviewTurnInput {
  /** The full transcript of the interview so far, oldest first. */
  messages: InterviewMessage[];
  /** When true, instructs the advisor to produce a final recommendation now. */
  forceFinalize?: boolean;
  /** Optional pre-filled student profile context to inject into the AI system prompt. */
  profileContext?: string;
  /** Optional session ID to associate the completed interview record with this session. */
  sessionId?: string;
}

export interface InterviewFinalizeInput {
  /** The full transcript of the interview so far, oldest first. */
  messages: InterviewMessage[];
  /** Optional session ID to link the finalized record to the active session. */
  sessionId?: string;
}

/**
 * Subtle, non-checklist progress signal (0-100).
 */
export interface InterviewProgress {
  /**
   * @minimum 0
   * @maximum 100
   */
  percent: number;
  /** A short human-readable stage label. */
  stage: string;
}

export interface InterviewRecommendation {
  recommendedMajor: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  matchScore: number;
  whyItFits: string[];
  alternativeMajors: string[];
  academicStrengths: string[];
  careerAdvice: string[];
  closingMessage: string;
  /**
   * Optional note about AAUP admission requirements and GPA eligibility.
   * @nullable
   */
  admissionNote?: string | null;
}

export type InterviewTurnKind =
  (typeof InterviewTurnKind)[keyof typeof InterviewTurnKind];

export const InterviewTurnKind = {
  question: "question",
  result: "result",
} as const;

export interface InterviewTurn {
  kind: InterviewTurnKind;
  /**
   * The next AI question for the student. Present when kind=question.
   * @nullable
   */
  question?: string | null;
  progress: InterviewProgress;
  /** Final recommendation. Present when kind=result. */
  recommendation?: InterviewRecommendation | null;
}

export interface InterviewError {
  error: string;
}

/**
 * @nullable
 */
export type InterviewResultRecordUser = {
  id?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
} | null;

/**
 * A completed interview record stored in the database.
 */
export interface InterviewResultRecord {
  /**
   * The session ID this record is linked to.
   * @nullable
   */
  sessionId?: string | null;
  recommendation?: InterviewRecommendation;
  fullConversation?: InterviewMessage[];
  savedAt?: string;
  /** @nullable */
  user?: InterviewResultRecordUser;
}

/**
 * Optional profile form data captured at each step.
 */
export type InterviewSessionDataProfileData = { [key: string]: unknown } | null;

export interface InterviewSessionData {
  id: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  title?: string | null;
  /** Optional profile form data captured at each step. */
  profileData?: InterviewSessionDataProfileData;
  messages: InterviewMessage[];
  progress?: InterviewProgress | null;
  recommendation?: InterviewRecommendation | null;
}

export interface InterviewSessionList {
  sessions: InterviewSessionData[];
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};