# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interview/start:
    post:
      operationId: startInterview
      tags: [interview]
      summary: Start a new AI interview
      description: |
        Returns the AI advisor's opening greeting and first question. The client should
        treat the returned turn as the first advisor message in a fresh transcript.
      responses:
        "200":
          description: The opening interview turn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewTurn"
        "500":
          description: AI service failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
  /interview/turn:
    post:
      operationId: interviewTurn
      tags: [interview]
      summary: Advance the AI interview by one turn
      description: |
        Submits the full transcript so far (messages between the student and the AI advisor)
        and returns either the next AI question or the final recommendation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewTurnInput"
      responses:
        "200":
          description: The next interview turn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewTurn"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
        "500":
          description: AI service failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
  /interview/finalize:
    post:
      operationId: finalizeInterview
      tags: [interview]
      summary: Force the AI to produce the final recommendation now
      description: |
        Submits the full transcript so far and asks the AI advisor to produce the final
        recommendation immediately, regardless of whether it would normally have done so.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewFinalizeInput"
      responses:
        "200":
          description: The final interview turn (kind=result)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewTurn"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
        "500":
          description: AI service failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
...
components:
...
  schemas:
...
    InterviewMessage:
      type: object
      properties:
        role:
          type: string
          enum: [student, advisor]
        content:
          type: string
      required:
        - role
        - content
    InterviewTurnInput:
      type: object
      properties:
        messages:
          type: array
          description: The full transcript of the interview so far, oldest first.
          items:
            $ref: "#/components/schemas/InterviewMessage"
        forceFinalize:
          type: boolean
          description: When true, instructs the advisor to produce a final recommendation now.
        profileContext:
          type: string
          description: Optional pre-filled student profile context to inject into the AI system prompt.
        sessionId:
          type: string
          description: Optional session ID to associate the completed interview record with this session.
      required:
        - messages
    InterviewFinalizeInput:
      type: object
      properties:
        messages:
          type: array
          description: The full transcript of the interview so far, oldest first.
          items:
            $ref: "#/components/schemas/InterviewMessage"
        sessionId:
          type: string
          description: Optional session ID to link the finalized record to the active session.
      required:
        - messages
    InterviewProgress:
      type: object
      description: Subtle, non-checklist progress signal (0-100).
      properties:
        percent:
          type: integer
          minimum: 0
          maximum: 100
        stage:
          type: string
          description: A short human-readable stage label.
      required:
        - percent
        - stage
    InterviewRecommendation:
      type: object
      properties:
        recommendedMajor:
          type: string
        matchScore:
          type: integer
          minimum: 0
          maximum: 100
        whyItFits:
          type: array
          items:
            type: string
        alternativeMajors:
          type: array
          items:
            type: string
        academicStrengths:
          type: array
          items:
            type: string
        careerAdvice:
          type: array
          items:
            type: string
        closingMessage:
          type: string
        admissionNote:
          type: ["string", "null"]
          description: Optional note about AAUP admission requirements and GPA eligibility.
      required:
        - recommendedMajor
        - matchScore
        - whyItFits
        - alternativeMajors
        - academicStrengths
        - careerAdvice
        - closingMessage
    InterviewTurn:
      type: object
      properties:
        kind:
          type: string
          enum: [question, result]
        question:
          type: ["string", "null"]
          description: The next AI question for the student. Present when kind=question.
        progress:
          $ref: "#/components/schemas/InterviewProgress"
        recommendation:
          oneOf:
            - $ref: "#/components/schemas/InterviewRecommendation"
            - type: "null"
          description: Final recommendation. Present when kind=result.
      required:
        - kind
        - progress
    InterviewError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...