# lib/api-zod/src/generated/types/interviewFinalizeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewMessage } from "./interviewMessage";

export interface InterviewFinalizeInput {
  /** The full transcript of the interview so far, oldest first. */
  messages: InterviewMessage[];
  /** Optional session ID to link the finalized record to the active session. */
  sessionId?: string;
}