# lib/api-zod/src/generated/types/interviewRecommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewRecommendation {
  recommendedMajor: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  matchScore: number;
  whyItFits: string[];
  alternativeMajors: string[];
  academicStrengths: string[];
  careerAdvice: string[];
  closingMessage: string;
  /**
   * Optional note about AAUP admission requirements and GPA eligibility.
   * @nullable
   */
  admissionNote?: string | null;
}