# lib/api-zod/src/generated/types/interviewTurnInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewMessage } from "./interviewMessage";

export interface InterviewTurnInput {
  /** The full transcript of the interview so far, oldest first. */
  messages: InterviewMessage[];
  /** When true, instructs the advisor to produce a final recommendation now. */
  forceFinalize?: boolean;
  /** Optional pre-filled student profile context to inject into the AI system prompt. */
  profileContext?: string;
  /** Optional session ID to associate the completed interview record with this session. */
  sessionId?: string;
}